requirements.txt
SQLAlchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
Faker>=37.0
prompt-toolkit>=3.0
python-dotenv>=1.0
pytest==8.4.1
click>=8.1